+++ packages/webpack-files/svg.js
@@ -1,7 +1,7 @@
 module.exports = {
   rules: [
     {
-      test: /\.svg$/,
+      test: /\.(svg)(\?v=[0-9]\.[0-9]\.[0-9])?$/,
       loader: 'file-loader'
     }
   ]