+++ packages/webpack-files/svg.js
@@ -1,0 +1,8 @@
+module.exports = {
+  rules: [
+    {
+      test: /\.(svg)(\?v=[0-9]\.[0-9]\.[0-9])?$/,
+      loader: 'file-loader'
+    }
+  ]
+};