+++ packages/webpack-files/yaml.js
@@ -1,0 +1,8 @@
+module.exports = {
+  rules: [
+    {
+      test: /\.yaml$/,
+      loader: 'yaml-loader'
+    }
+  ]
+};