+++ packages/webpack-less/index.js
@@ -5,6 +5,7 @@ module.exports = {
       use: [
         'style-loader',
         'css-loader',
+        'postcss-loader',
         'less-loader'
       ]
     }