+++ packages/webpack-files/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added JSON and YAML loaders.
 - Moved SVG loader from `fonts` loaders file to a new `svg` loader file.
+- Removed SVG files loader from the main exported file.
 
 # 1.1.0