+++ packages/webpack-files/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.2.0
+
+- Added JSON and YAML loaders.
+- Moved SVG loader from `fonts` loaders file to a new `svg` loader file.
+- Removed SVG files loader from the main exported file.
+
 # 1.1.0
 
 - Updated copy-webpack-plugin, file-loader and url-loader.